server/requirements.txt
Django==1.8.3
django-cors-headers==1.1.0
django-filter==0.11.0
djangorestframework==3.2.2
djangorestframework-filters==0.4.0
psycopg2==2.5.5
wheel==0.24.0
